linked links, modified things.
Linked up all the links, modified the footer to erase any links that had no page, changed the rss icon to github and linked to repo. Changed the My Budget button on the top to a dollar sign icon, added the logo. Changed the faq to reflect the changes made to the budget app.